Moved nap_conetent.go logic in the the agent sdk
